fix(vstsplittingcomponent): vstNodes are cleaned after destruction
ie references to nodes which have had half or all of their children removed, are fast-forwarded through or removed themself (respectively)

im actually not sure if this is definitely a fix. I think the added code is doing _something_ different and logically it feels more correct. but idk if the destruction still duplicates nodes or adds mass unphysically somehow.