feat: complete refactor, removed ReFrag() and CleanVST() functions and replaced them with the Orphan() function
i.e. the aim is to do something theoretically equivalent to what we were trying to do before, but on the fly without recursing through the VST multiple times. instead we update the VST as we go, i.e. as we mark fragments to be removed

also removed some deprecated / commented out code