feat: PlotSitesRandom now generates uniformly random points across the volumes it is testing
(or at least I think so lmao...)

before, the code used a gaussian dist with a set deviation which definitely wasnt uniform. But idk if this version is that different (superficially it looks to fragment very similarly).

performance wise i expect the time to find a valid site is more consistent (as its now independent of fragment size compared to before) but i haven't thoroughly tested this.